feat(auth): add modular SAML handler
