+++ README.md
@@ -59,9 +59,10 @@ This creates a volume named `pda-data` to persist the default SQLite database wi
    Other environment variables are mentioned in
    the [AppSettings.defaults](https://github.com/PowerDNS-Admin/PowerDNS-Admin/blob/master/powerdnsadmin/lib/settings.py) dictionary.
    To use a Docker-style secrets convention, one may append `_FILE` to the environment variables with a path to a file
-   containing the intended value of the variable (e.g. `SQLALCHEMY_DATABASE_URI_FILE=/run/secrets/db_uri`).   
+   containing the intended value of the variable (e.g. `SQLALCHEMY_DATABASE_URI_FILE=/run/secrets/db_uri`).
    Make sure to set the environment variable `SECRET_KEY` to a long, random
    string (https://flask.palletsprojects.com/en/1.1.x/config/#SECRET_KEY)
+   Optionally set `SAML_IMPLEMENTATION=modern` to enable the new pysaml2 based handler.
 
 2. Start docker container
    ```
