+++ powerdnsadmin/default_config.py
@@ -14,6 +14,7 @@ PORT = 9191
 SALT = '$2b$12$yLUMTIfl21FKJQpTkRQXCu'
 SAML_ASSERTION_ENCRYPTED = True
 SAML_ENABLED = False
+SAML_IMPLEMENTATION = os.getenv('SAML_IMPLEMENTATION', 'legacy')
 SECRET_KEY = 'e951e5a1f4b94151b360f47edf596dd2'
 SERVER_EXTERNAL_SSL = os.getenv('SERVER_EXTERNAL_SSL', True)
 SESSION_COOKIE_SAMESITE = 'Lax'
