+++ requirements.txt
@@ -34,6 +34,7 @@ pyotp==2.8.0
 pytest==7.4.4
 python-ldap==3.4.3
 python3-saml==1.15.0
+pysaml2>=7.0.2
 pytimeparse==1.1.8
 pytz==2022.7.1
 qrcode==7.3.1
